+++ packages/antd-management-fast-design-playground/src/components/displayer/FileViewer/index.jsx
@@ -62,7 +62,6 @@ class FileViewer extends PureComponent {
       showUrl,
       splitHeight,
       list,
-      menuZIndex,
       dataTransfer,
       onUploadButtonClick,
       onItemClick,
@@ -97,6 +96,7 @@ class FileViewer extends PureComponent {
         text: '删除附件',
         confirm: true,
         title: '将要删除附件，确定吗？',
+        placement: 'bottomRight',
       });
     }
 
@@ -130,7 +130,6 @@ class FileViewer extends PureComponent {
             size: 'small',
             text: '查看',
             icon: iconBuilder.read(),
-            itemPanelZIndex: menuZIndex,
             handleButtonClick: ({ handleData }) => {
               onItemClick(handleData);
             },
@@ -209,7 +208,6 @@ FileViewer.defaultProps = {
   canRemove: false,
   showUrl: false,
   splitHeight: 0,
-  menuZIndex: null,
   dataTransfer: (o) => o,
   onItemClick: null,
   onUploadButtonClick: null,