+++ packages/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 1.8.169
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.103
+  - antd-management-fast-component@2.6.141
+
 ## 1.8.168
 
 ### Patch Changes